+++ src/main/java/org/kdb/inside/brains/psi/refs/QBaseReference.java
@@ -33,7 +33,7 @@ public abstract class QBaseReference<T extends QPsiElement> extends PsiPolyVaria
 
     @Override
     public ResolveResult @NotNull [] multiResolve(boolean incompleteCode) {
-        return resolveElement(myElement);
+        return resolveGlobalDeclaration(myElement);
     }
 
     @Override
@@ -74,7 +74,7 @@ public abstract class QBaseReference<T extends QPsiElement> extends PsiPolyVaria
         return false;
     }
 
-    protected ResolveResult[] resolveElement(T element) {
+    protected ResolveResult[] resolveGlobalDeclaration(T element) {
         final PsiFile file = element.getContainingFile();
         if (file == null) {
             return ResolveResult.EMPTY_ARRAY;
@@ -90,7 +90,7 @@ public abstract class QBaseReference<T extends QPsiElement> extends PsiPolyVaria
         if (initial == null) {
             final GlobalSearchScope scope = GlobalSearchScope.allScope(element.getProject());
             final Collection<DeclarationRef> declarations = QIndexService.getInstance(element).getDeclarations(name, scope);
-            return multi(declarations.stream().filter(DeclarationRef::isGlobalDeclaration));
+            return multi(declarations.stream().filter(DeclarationRef::isGlobal));
         }
         return single(initial);
     }